+++ core/src/main/java/io/code/crm/init/SetupService.java
@@ -1,0 +1,16 @@
+package io.code.crm.init;
+
+import org.springframework.stereotype.Service;
+
+import lombok.extern.java.Log;
+
+@Log
+@Service
+public class SetupService {
+
+    public void execute() {
+        log.info("CRM Setup");
+
+    }
+
+}
