+++ core/src/main/java/io/code/crm/model/contact/ContactNotFoundException.java
@@ -1,0 +1,14 @@
+package io.code.crm.model.contact;
+
+public class ContactNotFoundException extends Exception {
+
+    /**
+     * 
+     */
+    private static final long serialVersionUID = 1L;
+
+    public ContactNotFoundException(String errorMessage) {
+        super(errorMessage);
+    }
+
+}
