+++ core/src/main/java/io/code/crm/service/ContactRepository.java
@@ -1,0 +1,10 @@
+package io.code.crm.service;
+
+import java.util.UUID;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import io.code.crm.model.contact.Contact;
+
+public interface ContactRepository extends JpaRepository<Contact, UUID> {
+}
