+++ core/src/main/resources/data.sql
@@ -1,0 +1,3 @@
+
+
+INSERT INTO contact (id, name, first_name, surname1, surname2, created_time, updated_time) VALUES (RANDOM_UUID(), 'John Smith', 'John', 'Smith', 'Matrix',  CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP());
