+++ rest/src/main/java/io/code/crm/rest/RestApplication.java
@@ -1,0 +1,13 @@
+package io.code.crm.rest;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RestApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RestApplication.class, args);
+	}
+
+}
