+++ rest/src/main/java/io/code/crm/rest/contact/rest/config/LoadCrmCore.java
@@ -1,0 +1,10 @@
+package io.code.crm.rest.contact.rest.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = "io.code.crm.service")
+public class LoadCrmCore {
+    
+}
