+++ rest/src/test/java/io/code/crm/rest/RestApplicationTests.java
@@ -1,0 +1,34 @@
+package io.code.crm.rest;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.List;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import io.code.crm.model.contact.Contact;
+import io.code.crm.service.ContactServiceFull;
+
+
+@SpringBootTest
+class RestApplicationTests {
+
+
+
+
+	@Autowired
+	ContactServiceFull contactServiceFull;
+	
+	@Test
+	void testGetAllContacts(){
+
+		List<Contact> contactList = this.contactServiceFull.getContactList();
+		for(Contact cc : contactList){
+			System.out.println(cc.toString());
+		}
+		assertTrue(contactList!=null && !contactList.isEmpty());
+	}
+
+}
