--- src/main/resources/data.sql
@@ -1,3 +1,0 @@
-
-
-INSERT INTO contact (id, name, first_name, surname1, surname2, created_time, updated_time) VALUES (RANDOM_UUID(), 'Diego Pérez', 'Diego', 'Pérez', 'Pastor',  CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP());
