--- src/test/java/io/code/CrmApplicationTests.java
@@ -1,13 +1,0 @@
-package io.code;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CrmApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
